evals/acme_doc_assistant/app/requirements.txt
flask>=3.0
google-auth>=2.0
google-auth-oauthlib>=1.0
google-api-python-client>=2.0
google-cloud-storage>=2.0
google-cloud-secret-manager>=2.0
gunicorn>=22.0
